+++ aftermath-core/web/src/lib.rs
@@ -2,7 +2,7 @@ mod utils;
 
 use input_tree::{input_node::InputNode, row::InputRow};
 use parser::{parse_rules::ParserRules, ParseError, ParseResult, SyntaxNode};
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use utils::set_panic_hook;
 use wasm_bindgen::prelude::*;
 
@@ -36,6 +36,12 @@ struct MathParseResult {
 pub struct MathParser {
     parser_rules: ParserRules<'static>,
     serializer: serde_wasm_bindgen::Serializer,
+}
+
+#[derive(Serialize, Deserialize)]
+/// Only for temporary copy-pasting
+struct JsonSerializedFormula {
+    input: InputRow,
 }
 
 #[wasm_bindgen]
@@ -76,6 +82,23 @@ impl MathParser {
     pub fn get_token_names(&self) -> Result<JsValue, JsValue> {
         let token_names = self.parser_rules.get_token_names();
         let serialized_result = token_names.serialize(&self.serializer)?;
+        Ok(serialized_result)
+    }
+
+    pub fn serialize(&self, layout_row: JsValue) -> Result<String, JsValue> {
+        let layout: InputRow = serde_wasm_bindgen::from_value(layout_row)?;
+        let value = JsonSerializedFormula { input: layout };
+
+        let json_string = serde_json::to_string(&value)
+            .map_err(|e| serde_wasm_bindgen::Error::new(e.to_string()))?;
+        Ok(json_string)
+    }
+
+    pub fn deserialize(&self, json_string: String) -> Result<JsValue, JsValue> {
+        let value: InputRow = serde_json::from_str(&json_string)
+            .map(|v: JsonSerializedFormula| v.input)
+            .map_err(|e| serde_wasm_bindgen::Error::new(e.to_string()))?;
+        let serialized_result = value.serialize(&self.serializer)?;
         Ok(serialized_result)
     }
 }