+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# next
+
+- Copy-pasting
+
 # v0.2.3
 
 - Add table selections